Use long for graph nodeId - Part 2
Changes NeighborQueue to use a priority queue instead of LongHeap.

+++ lucene/core/src/java/org/apache/lucene/util/hnsw/HnswGraphSearcher.java
@@ -189,7 +189,7 @@ public class HnswGraphSearcher {
       KnnCollector results,
       RandomVectorScorer scorer,
       int level,
-      final int[] eps,
+      final long[] eps,
       HnswGraph graph,
       Bits acceptOrds)
       throws IOException {
